+++ tests/integ/test_pytorch_train.py
@@ -37,6 +37,7 @@ def fixture_training_job(sagemaker_session, pytorch_full_version):
         return pytorch.latest_training_job.name
 
 
+@pytest.mark.continuous_testing
 def test_sync_fit_deploy(pytorch_training_job, sagemaker_session):
     # TODO: add tests against local mode when it's ready to be used
     endpoint_name = 'test-pytorch-sync-fit-attach-deploy{}'.format(sagemaker_timestamp())